feat : 관리자단 - 신고 전체 리스트 조회 api

+++ src/main/java/com/dalbong/cafein/dto/admin/report/AdminReportListResDto.java
@@ -13,7 +13,7 @@ import java.util.List;
 @Data
 public class AdminReportListResDto {
 
-    private int reportCnt;
+    private long reportCnt;
 
     private PageResultDTO<AdminReportResDto, Object[]> reportResDtoList;
 }
